+++ projects/ng-translation/src/lib/default-translation.converter.ts
@@ -1,0 +1,16 @@
+/* 3rd party libraries */
+
+/* locally accessible feature module code, always use relative path */
+import { Resource, TranslationConverter } from './models';
+
+export class DefaultTranslationConverter implements TranslationConverter {
+
+  convert(
+    language: string,
+    resource: Resource,
+    translations: any
+  ): object {
+
+    return typeof translations === 'object' ? translations : { };
+  }
+}