+++ projects/ng-translation/src/lib/directives/index.ts
@@ -1,0 +1,2 @@
+export * from './translate.directive';
+export * from './translate-params.directive';