+++ projects/ng-translation/src/lib/models/transpile-extender.model.ts
@@ -1,0 +1,17 @@
+/* 3rd party libraries */
+import { InjectionToken } from '@angular/core';
+
+/* locally accessible feature module code, always use relative path */
+import { FormatData } from './format-data.model';
+
+export interface TranspileExtender {
+
+  transpile(
+    format: string,
+    data: FormatData
+  ): string | undefined;
+}
+
+export const NGT_TRANSPILE_EXTENDER = new InjectionToken<TranspileExtender>(
+  'NGT_TRANSPILE_EXTENDER'
+);