+++ projects/ng-translation/src/lib/services/index.ts
@@ -1,0 +1,4 @@
+export * from './localization.service';
+export * from './messenger.service';
+export * from './translation.service';
+export * from './transpiler.service';