+++ projects/ng-translation/src/lib/translatable-option.model.ts
@@ -1,6 +1,0 @@
-export interface TranslatableOption {
-
-  value: string;
-  text: string;
-  selected: boolean;
-}