+++ projects/ng-translation/src/lib/translation-config.model.ts
@@ -1,7 +1,0 @@
-export class TranslationConfig {
-
-  translationUrl: string;
-  sections: Array<string>;
-  defaultLanguage: string;
-  disableWarnings?: boolean;
-}