+++ projects/test-web-site/src/app/winter/index/index.component.ts
@@ -1,14 +1,9 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
   selector: 'app-index',
   templateUrl: './index.component.html',
   styleUrls: ['./index.component.css'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  // changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class IndexComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() { }
-}
+export class IndexComponent { }